Fix File Encoding, ViewAttributes for cyrillic
Some lazy and dirty fixes on strings extracting from database. Some
strings like file content are treated fine, but file attributes and file
encoding views was shown corrupted. Strings extracting like they are
UTF-8, but in reality they are 1251 (Win-Russian). Looks like it is
independent from which encoding set in global options in R. It is
obvious for me (thou it is not prove that i am right) that similar
problems should be experienced with other charsets except English. But
since i use this pkg for my own work, such fixes are enough for now.